Remove unnecessary warning on increment/decrement.
When trying a increment/decrement with create = 0 (which is the default),
not only false is returned, but also a warning is shown. This is not
needed since false is enough information that the key doesn't exist. Also,
the warning is hard to catch when working from a user-level.

This change makes sure that only this specific information is hidden and
all other possible errors are still shown to the user.

Change-Id: Idd73d407cb46b8d4fb939ab844814ac5f68ed186
Reviewed-on: http://review.couchbase.org/20259
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Reviewed-by: Mordechai Nunberg <mnunberg@haskalah.org>
Tested-by: Mordechai Nunberg <mnunberg@haskalah.org>