version() fix, add clientVersion() -- PCBC-108
This changeset addresses PCBC-108:
* version() now returns the server's version as determined by the
libcouchbase version function.  (Notably this is not the release
version name of the server, but rather more like the memcached
protocol level.  The release version name of the server may be
determined by calling the existing getStats()/couchbase_get_stats
functions and checking the "ep_version" key of the returned data.)
* a "get client version" function/method pair is added that acts
like the previous version functionality, in that it returns the
version of the couchbase extension itself, so that no data availability
is lost for the extension users.

Change-Id: Ia37466f48c50ede86b8c0c7e742f80ec2c1bd37e
Reviewed-on: http://review.couchbase.org/20862
Reviewed-by: Mordechai Nunberg <mnunberg@haskalah.org>
Tested-by: Mordechai Nunberg <mnunberg@haskalah.org>